refactor(console): relocalte the component files
relocate the component files
