+++ packages/console/src/pages/CustomizeJwt/CustomizerItem/index.module.scss
